pyproject.toml
[project]
name = "torch-experiment"
version = "0.1.0"
description = ""
authors = [
    { name = "yuiyeong", email = "joyuiyeong@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "numpy (>=2.3.1,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "wandb (>=0.21.0,<0.22.0)",
    "torch (>=2.7.1,<2.8.0)",
    "lightning (>=2.5.2,<2.6.0)",
    "hydra-core (>=1.3.2,<1.4.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "tokenizers (>=0.21.2,<1.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "peft (>=0.16.0,<0.17.0)",
    "accelerate (>=1.9.0,<2.0.0)",
    "deepspeed (>=0.17.2,<0.18.0)",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"
pre-commit = "^4.2.0"
jupyterlab = "^4.4.5"
notebook = "^7.4.4"
ipykernel = "^6.29.5"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
